riscv32-noop,blinky_demo: fix unmatched size of queue entry
* under rv64, sizeof(unsign long) = 8 != sizeof(uint32_t)

+++ Demo/riscv64-noop/blinky_demo/main_blinky.c
@@ -113,7 +113,7 @@ static QueueHandle_t xQueue = NULL;
 void main_blinky( void )
 {
 	/* Create the queue. */
-	xQueue = xQueueCreate( mainQUEUE_LENGTH, sizeof( uint32_t ) );
+	xQueue = xQueueCreate( mainQUEUE_LENGTH, sizeof( unsigned long ) );
 
 	if( xQueue != NULL )
 	{
